Bug upon License update only fixed
Seemed to be a date issue. Try catch has been changed to be more descriptive upon license update only. Code could be cleaner as it now has a few redundant lines but <3 I don't care. Unless you are a teacher or PO, then I do care and will fix this in the future but I have forgotten<3